feat: 점수 최댓값 초과 여부 확인 기능 구현
