+++ README.md
@@ -62,6 +62,7 @@ graph TD
 - [x] 숫자를 가진다.
 - [x] 점수를 계산한다.
 - [x] 에이스인지 확인한다.
+- [x] 점수 최댓값 초과 여부를 확인한다.
 
 ### 덱
